Fix tests and allow firestore connection
