+++ src/lib/backend/firebase/deck/index.spec.js
@@ -6,7 +6,7 @@ import Deck from './index';
 describe('Decks', () => {
 	test('getDecks resolve a promise with a firebase collection if the are not problems with the request', (done) => {
 		const dbMock = {
-			collections: () => {
+			collection: () => {
 				return {
 					get: () => Promise.resolve([ { _id: 'base '}])
 				}
@@ -21,7 +21,7 @@ describe('Decks', () => {
 
 	test('getDecks rejects a promise if the are problems with the request', (done) => {
 		const dbMock = {
-			collections: () => {
+			collection: () => {
 				return {
 					get: () => Promise.reject(new Error('cannot get the decks!'))
 				}
