--- src/lib/backend/firebase/room/index.spec.js
@@ -1,13 +1,0 @@
-/* src/lib/backend/firebase/room/index.spec.js */
-'use strict';
-
-import Room from './index';
-
-describe('room', () => {
-	test('createRoom reject the promise if createdBy is invalid', (done) => {
-		Room.createRoom({}, null).catch((error) => {
-			expect(error).toBeInstanceOf(TypeError);
-			done();
-		});
-	});
-});
